Moved user.service to correct folder
